Fixes for bonus and summary tables
adjusted the end of day sales logic so it just takes the cash and credit into account as it was previously double counting when it got to the summary screen. Added unique ID into sql logic for storeID for end of day sales for filtering purposes when it goes to summary table. Added filter dropdown menu into summary table to see total sales, profit, etc for each store as well as all stores together. Prompt user with warning message if someone already input an end of day sale for the day before overriding. Adjusted readme.